+++ src/Addons/34thPRC_AceArsenalExtended/data/aax_infos/config.hpp
@@ -1,5 +1,6 @@
 class cfgWeapons
 {
 	#include "weapons\config.hpp"
+	#include "vests\config.hpp"
 };