+++ src/Addons/34thPRC_AceArsenalExtended/data/aax_infos/vests/ma_vests/config.hpp
@@ -1,0 +1,316 @@
+class MA_Marine_Pouches;
+class Marine_Pouches_Var0
+{
+	model = "MA_Marine_Pouches";
+	basic = "va0";
+	rifle = "";
+	tl = "";
+	med = "";
+	eng = "";
+};
+class Marine_Pouches_Var1
+{
+	model = "MA_Marine_Pouches";
+	basic = "va1";
+	rifle = "";
+	tl = "";
+	med = "";
+	eng = "";
+};
+class Marine_Pouches_Var2
+{
+	model = "MA_Marine_Pouches";
+	basic = "va2";
+	rifle = "";
+	tl = "";
+	med = "";
+	eng = "";
+};
+class Marine_Pouches_Var3
+{
+	model = "MA_Marine_Pouches";
+	basic = "va3";
+	rifle = "";
+	tl = "";
+	med = "";
+	eng = "";
+};
+class Marine_Pouches_Var4
+{
+	model = "MA_Marine_Pouches";
+	basic = "va4";
+	rifle = "";
+	tl = "";
+	med = "";
+	eng = "";
+};
+class MA_Marine_Pouches_Engineer_Heavy_A
+{
+	model = "MA_Marine_Pouches";
+	basic = "";
+	rifle = "";
+	tl = "";
+	med = "";
+	eng = "eha";
+};
+class MA_Marine_Pouches_Engineer_Heavy_B
+{
+	model = "MA_Marine_Pouches";
+	basic = "";
+	rifle = "";
+	tl = "";
+	med = "";
+	eng = "ehb";
+};
+class MA_Marine_Pouches_Engineer_Medium_A
+{
+	model = "MA_Marine_Pouches";
+	basic = "";
+	rifle = "";
+	tl = "";
+	med = "";
+	eng = "ema";
+};
+class MA_Marine_Pouches_Engineer_Medium_B
+{
+	model = "MA_Marine_Pouches";
+	basic = "";
+	rifle = "";
+	tl = "";
+	med = "";
+	eng = "emb";
+};
+class MA_Marine_Pouches_Engineer_Light_A
+{
+	model = "MA_Marine_Pouches";
+	basic = "";
+	rifle = "";
+	tl = "";
+	med = "";
+	eng = "ela";
+};
+class MA_Marine_Pouches_Engineer_Light_B
+{
+	model = "MA_Marine_Pouches";
+	basic = "";
+	rifle = "";
+	tl = "";
+	med = "";
+	eng = "elb";
+};
+class MA_Marine_Pouches_Medic_Heavy_A
+{
+	model = "MA_Marine_Pouches";
+	basic = "";
+	rifle = "";
+	tl = "";
+	med = "mha";
+	eng = "";
+};
+class MA_Marine_Pouches_Medic_Heavy_B
+{
+	model = "MA_Marine_Pouches";
+	basic = "";
+	rifle = "";
+	tl = "";
+	med = "mhb";
+	eng = "";
+};
+class MA_Marine_Pouches_Medic_Medium_A
+{
+	model = "MA_Marine_Pouches";
+	basic = "";
+	rifle = "";
+	tl = "";
+	med = "mma";
+	eng = "";
+};
+class MA_Marine_Pouches_Medic_Medium_B
+{
+	model = "MA_Marine_Pouches";
+	basic = "";
+	rifle = "";
+	tl = "";
+	med = "mmb";
+	eng = "";
+};
+class MA_Marine_Pouches_Medic_Light_A
+{
+	model = "MA_Marine_Pouches";
+	basic = "";
+	rifle = "";
+	tl = "";
+	med = "mla";
+	eng = "";
+};
+class MA_Marine_Pouches_Medic_Light_B
+{
+	model = "MA_Marine_Pouches";
+	basic = "";
+	rifle = "";
+	tl = "";
+	med = "mlb";
+	eng = "";
+};
+class MA_Marine_Pouches_Teamleader_Heavy_A
+{
+	model = "MA_Marine_Pouches";
+	basic = "";
+	rifle = "";
+	tl = "tha";
+	med = "";
+	eng = "";
+};
+class MA_Marine_Pouches_Teamleader_Heavy_B
+{
+	model = "MA_Marine_Pouches";
+	basic = "";
+	rifle = "";
+	tl = "thb";
+	med = "";
+	eng = "";
+};
+class MA_Marine_Pouches_Teamleader_Medium_A
+{
+	model = "MA_Marine_Pouches";
+	basic = "";
+	rifle = "";
+	tl = "tma";
+	med = "";
+	eng = "";
+};
+class MA_Marine_Pouches_Teamleader_Medium_B
+{
+	model = "MA_Marine_Pouches";
+	basic = "";
+	rifle = "";
+	tl = "tmb";
+	med = "";
+	eng = "";
+};
+class MA_Marine_Pouches_Teamleader_Light_A
+{
+	model = "MA_Marine_Pouches";
+	basic = "";
+	rifle = "";
+	tl = "tla";
+	med = "";
+	eng = "";
+};
+class MA_Marine_Pouches_Teamleader_Light_B
+{
+	model = "MA_Marine_Pouches";
+	basic = "";
+	rifle = "";
+	tl = "tlb";
+	med = "";
+	eng = "";
+};
+class MA_Marine_Pouches_Rifleman_Heavy_A
+{
+	model = "MA_Marine_Pouches";
+	basic = "";
+	rifle = "rha";
+	tl = "";
+	med = "";
+	eng = "";
+};
+class MA_Marine_Pouches_Rifleman_Heavy_B
+{
+	model = "MA_Marine_Pouches";
+	basic = "";
+	rifle = "rhb";
+	tl = "";
+	med = "";
+	eng = "";
+};
+class MA_Marine_Pouches_Rifleman_Heavy_C
+{
+	model = "MA_Marine_Pouches";
+	basic = "";
+	rifle = "rhc";
+	tl = "";
+	med = "";
+	eng = "";
+};
+class MA_Marine_Pouches_Rifleman_Heavy_D
+{
+	model = "MA_Marine_Pouches";
+	basic = "";
+	rifle = "rhd";
+	tl = "";
+	med = "";
+	eng = "";
+};
+class MA_Marine_Pouches_Rifleman_Medium_A
+{
+	model = "MA_Marine_Pouches";
+	basic = "";
+	rifle = "rma";
+	tl = "";
+	med = "";
+	eng = "";
+};
+class MA_Marine_Pouches_Rifleman_Medium_B
+{
+	model = "MA_Marine_Pouches";
+	basic = "";
+	rifle = "rmb";
+	tl = "";
+	med = "";
+	eng = "";
+};
+class MA_Marine_Pouches_Rifleman_Medium_C
+{
+	model = "MA_Marine_Pouches";
+	basic = "";
+	rifle = "rmc";
+	tl = "";
+	med = "";
+	eng = "";
+};
+class MA_Marine_Pouches_Rifleman_Medium_D
+{
+	model = "MA_Marine_Pouches";
+	basic = "";
+	rifle = "rmd";
+	tl = "";
+	med = "";
+	eng = "";
+};
+class MA_Marine_Pouches_Rifleman_Light_A
+{
+	model = "MA_Marine_Pouches";
+	basic = "";
+	rifle = "rla";
+	tl = "";
+	med = "";
+	eng = "";
+};
+class MA_Marine_Pouches_Rifleman_Light_B
+{
+	model = "MA_Marine_Pouches";
+	basic = "";
+	rifle = "rlb";
+	tl = "";
+	med = "";
+	eng = "";
+};
+class MA_Marine_Pouches_Rifleman_Light_C
+{
+	model = "MA_Marine_Pouches";
+	basic = "";
+	rifle = "rlc";
+	tl = "";
+	med = "";
+	eng = "";
+};
+class MA_Marine_Pouches_Rifleman_Light_D
+{
+	model = "MA_Marine_Pouches";
+	basic = "";
+	rifle = "rld";
+	tl = "";
+	med = "";
+	eng = "";
+};