+++ src/Addons/34thPRC_AceArsenalExtended/data/aax_models/config.hpp
@@ -1,6 +1,7 @@
 class cfgWeapons
 {
 	#include "weapons\config.hpp"
+	#include "vests\config.hpp"
 };
 class cfgVehicles
 {