+++ src/Addons/34thPRC_AceArsenalExtended/data/aax_models/vests/config.hpp
@@ -1,0 +1,1 @@
+#include "ma_vests\config.hpp"